+++ src/component.js
@@ -26,7 +26,8 @@
         //     container: $element, // return the parent container associated with hash (#parent/hash/:id)
         //     parentNavMap: {} // Parent node in the navigation tree
         // }
-        _screenMetadata
+        _screenMetadata,
+        _debugMode
         ;
 
     
@@ -55,6 +56,9 @@
         _lastLoadedDependencies = [];
 
         _paths = [];
+
+        // By default debug is disabled
+        _debugMode = false;
 
         iris.on("iris-reset", function () {
             $(window).off("hashchange");
@@ -515,9 +519,10 @@
     }
 
 
-    var Component = function(id, $container, fileJs) {
+    var Component = function(id, $container, fileJs, type) {
         iris.Settable.call(this);
 
+        this.type = type;
         this.id = id;
         this.uis = []; // child UIs
         this.uisMap = {}; // UIs sorted by id
@@ -908,7 +913,7 @@
     // UI
     //
     var UI = function($container, id, fileJs, settings, tmplMode, parentUI) {
-        Component.call(this, id, $container, fileJs);
+        Component.call(this, id, $container, fileJs, 'ui');
 
         var jqToHash = _jqToHash($container);
 
@@ -942,7 +947,7 @@
     //
     var Screen = function(path) {
         var screenMeta = _screenMetadata[path];
-        Component.call(this, path, screenMeta.container, screenMeta.js);
+        Component.call(this, path, screenMeta.container, screenMeta.js, 'screen');
 
         this.params = {};
         this.screenConId = null;
@@ -1047,6 +1052,65 @@
         }
 
     }
+
+    //
+    // Debug mode
+    //
+    function _debug (enabled) {
+        var $doc = $(window.document);
+        if ( enabled ) {
+            $doc.on('keydown', _debugModeOnKeyDown);
+
+            var style = document.createElement('style');
+            style.type = 'text/css';
+            style.innerHTML = 
+             '.iris-debug-ui { outline: 2px dotted blue; }' +
+             '.iris-debug-ui-info { font-size: 12px; color: white; background-color: blue; }' +
+
+             '.iris-debug-screen { outline: 2px dotted red; }' +
+             '.iris-debug-screen-info { font-size: 12px; color: white; background-color: red; }';
+
+            document.getElementsByTagName('head')[0].appendChild(style);
+
+        } else {
+            $doc.off('keydown', _debugModeOnKeyDown);
+        }
+    }
+
+    function _debugModeOnKeyDown (e) {
+        // Control + Shift + Alt + D
+        if ( e.shiftKey && e.ctrlKey && e.altKey &&
+             e.keyCode !== 16 && e.keyCode === 68 ) {
+
+            _debugMode = !_debugMode;
+
+            var screen;
+            for ( var key in _screen ) {
+                screen = _screen[key];
+
+                _applyDebugMode(screen);
+                
+                for ( var f = 0, F = screen.uis.length; f < F; f++ ) {
+                    _applyDebugMode( screen.uis[f] );
+                }
+            }
+        }
+    }
+
+    function _applyDebugMode (component) {
+        component.template.toggleClass('iris-debug-' + component.type, _debugMode);
+        
+        if ( _debugMode ) {
+            // Add debug info label
+            component.debugElement = $(
+                '<span class="iris-debug-' + component.type + '-info">' +
+                   component.id + ' [' + component.fileJs + ',' + component.fileTmpl + ']' +
+                '</span>').prependTo(component.template);
+        } else {
+            // Remove debug info label
+            component.debugElement.remove();
+        }
+    }
     
     iris.screen = _registerScreen;
     iris.destroyScreen = _destroyScreenByPath;
@@ -1056,6 +1120,7 @@
     iris.tmpl = _registerTmpl;
     iris.resource = _registerRes;
     iris.include = _load;
+    iris.debug = _debug;
 
     //
     // Classes