+++ RELEASES.md
@@ -6,6 +6,7 @@ _Creation date: 2012-01-12_
   * upd|fix|new|dep|rmv - description
 
 [0.5.7-SNAPSHOT] 2014-XX-XX
+ * [new] Iris debug mode, Closed #169
 
 
 [0.5.6] 2014-01-23